added id to insert User into Blockchainuser table

+++ BlockchainProjekt/src/main/java/model/entity/Blockchainuser.java
@@ -1,14 +1,19 @@
 package model.entity;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.security.PublicKey;
 import java.util.Arrays;
+import java.util.Collection;
 import java.util.Date;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.EntityTransaction;
 import javax.persistence.Id;
 import javax.persistence.Table;
+
+import main.java.dao.NoSuchRowException;
 
 //create table blockchainuser (
 //id    	smallint primary key,
@@ -23,8 +28,8 @@ public class Blockchainuser {
 	private static final long serialVersionUID = 1L;
 
 	@Id
-	@Column(name = "id")
-	private int id;
+	@Column(name="id")
+	private BigDecimal id;
 
 	@Column(name = "userid")
 	private String userid;
@@ -48,11 +53,11 @@ public class Blockchainuser {
 		this.publickey = publicKey;
 	}
 
-	public int getId() {
+	public BigDecimal getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(BigDecimal id) {
 		this.id = id;
 	}
